backend/requirements.txt
anyio==4.11.0
attrs==25.4.0
base58==2.1.1
bcrypt==3.2.0
bidict==0.23.1
blinker==1.9.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
coverage==7.11.0
cryptography==44.0.0
Deprecated==1.3.1
dnspython==2.8.0
execnet==2.1.1
Flask==3.1.1
flask-cors==6.0.1
Flask-Limiter==3.5.0
Flask-Login==0.6.3
Flask-SocketIO==5.5.1
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
hvac==2.3.0
hypothesis==6.130.11
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.6.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
# motor==3.7.1
ordered-set==4.1.0
packaging==25.0
passlib==1.7.4
pluggy==1.6.0
pycparser==2.23
pyee==9.1.1
Pygments==2.19.2
PyJWT==2.3.0
pymongo==4.15.3
PyNaCl==1.6.0
pytest==8.3.5
pytest-cov==5.0.0
pytest-timeout==2.3.1
pytest-xdist==3.6.1
python-dotenv==1.1.1
python-engineio==4.12.3
python-socketio==5.14.1
redis==6.2.0
requests==2.32.4
# resilient-python-cache==0.1.1
rich==13.9.4
simple-websocket==1.1.0
simplejson==3.19.3
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
typing_extensions==4.15.0
urllib3==2.5.0
websockets==10.4
Werkzeug==3.1.3
wrapt==2.0.0
wsproto==1.2.0

# AI/ML dependencies for semantic search
torch>=2.0.0
open_clip_torch>=2.20.0
pillow>=10.0.0
qdrant-client>=1.7.0
numpy>=1.24.0
